chore(test): refactor build system test and adding interface for file struct output (#88)
