+++ backend/src/build-system/handlers/file-manager/file-structure/index.ts
@@ -1,4 +1,9 @@
-import { BuildHandler, BuildOpts, BuildResult } from 'src/build-system/types';
+import {
+  BuildHandler,
+  BuildOpts,
+  BuildResult,
+  FileStructOutput,
+} from 'src/build-system/types';
 import { BuilderContext } from 'src/build-system/context';
 import { prompts } from './prompt';
 import { Logger } from '@nestjs/common';
@@ -8,7 +13,7 @@ import { removeCodeBlockFences } from 'src/build-system/utils/strings';
  * FileStructureHandler is responsible for generating the project's file and folder structure
  * based on the provided documentation.
  */
-export class FileStructureHandler implements BuildHandler<string> {
+export class FileStructureHandler implements BuildHandler<FileStructOutput> {
   readonly id = 'op:FILE:STRUCT';
   private readonly logger: Logger = new Logger('FileStructureHandler');
 
@@ -21,7 +26,7 @@ export class FileStructureHandler implements BuildHandler<string> {
   async run(
     context: BuilderContext,
     opts?: BuildOpts,
-  ): Promise<BuildResult<string>> {
+  ): Promise<BuildResult<FileStructOutput>> {
     this.logger.log('Generating File Structure Document...');
 
     // Retrieve projectName from context
@@ -153,7 +158,10 @@ export class FileStructureHandler implements BuildHandler<string> {
 
     return {
       success: true,
-      data: removeCodeBlockFences(fileStructureJsonContent),
+      data: {
+        fileStructure: removeCodeBlockFences(fileStructureContent),
+        jsonFileStructure: removeCodeBlockFences(fileStructureJsonContent),
+      },
     };
   }
 }
